Reverse the bias so enforcement happens unless you go out of your way to avoid it.
