+++ clock_test.go
@@ -232,7 +232,7 @@ func TestClock_Timer_Reset(t *testing.T) {
 
 // Ensure reset can be called immediately after reading channel
 func TestClock_Timer_Reset_Unlock(t *testing.T) {
-	clock := NewMock()
+	clock := NewUnsynchronizedMock()
 	timer := clock.NewTimer(1 * time.Second)
 
 	var wg sync.WaitGroup
@@ -240,14 +240,10 @@ func TestClock_Timer_Reset_Unlock(t *testing.T) {
 	go func() {
 		defer wg.Done()
 
-		select {
-		case <-timer.C:
-			timer.Reset(1 * time.Second)
-		}
+		<-timer.C
+		timer.Reset(1 * time.Second)
 
-		select {
-		case <-timer.C:
-		}
+		<-timer.C
 	}()
 
 	clock.Add(2 * time.Second)
